# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Quiz Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface CustomQuestion {
  question: string;
  /**
   * @minItems 4
   * @maxItems 4
   */
  options: string[];
  correctAnswer: string;
}

export interface CreateRoomRequest {
  hostName: string;
  /**
   * @minimum 10
   * @maximum 100
   */
  questionCount: number;
  customQuestions?: CustomQuestion[];
}

export interface JoinRoomRequest {
  guestName: string;
}

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  waiting: "waiting",
  playing: "playing",
  finished: "finished",
} as const;

export interface Room {
  code: string;
  status: RoomStatus;
  hostName: string;
  guestName?: string | null;
  questionCount: number;
  customQuestions?: CustomQuestion[];
  createdAt: string;
}

export interface MatchmakingRequest {
  playerName: string;
  playerId: string;
}

export type MatchmakingResponseStatus =
  (typeof MatchmakingResponseStatus)[keyof typeof MatchmakingResponseStatus];

export const MatchmakingResponseStatus = {
  queued: "queued",
  matched: "matched",
} as const;

export interface MatchmakingResponse {
  status: MatchmakingResponseStatus;
  roomCode?: string | null;
  playerId: string;
}