# lib/api-zod/src/generated/types/createRoomRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Quiz Game API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomQuestion } from "./customQuestion";

export interface CreateRoomRequest {
  hostName: string;
  /**
   * @minimum 10
   * @maximum 100
   */
  questionCount: number;
  customQuestions?: CustomQuestion[];
}