# lib/api-zod/src/generated/types/customQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Quiz Game API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomQuestion {
  question: string;
  /**
   * @minItems 4
   * @maxItems 4
   */
  options: string[];
  correctAnswer: string;
}