# lib/api-zod/src/generated/types/matchmakingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Quiz Game API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchmakingRequest {
  playerName: string;
  playerId: string;
}