# lib/api-zod/src/generated/types/matchmakingResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Quiz Game API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchmakingResponseStatus } from "./matchmakingResponseStatus";

export interface MatchmakingResponse {
  status: MatchmakingResponseStatus;
  roomCode?: string | null;
  playerId: string;
}