# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Islamic Quiz Game API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomQuestion } from "./customQuestion";
import type { RoomStatus } from "./roomStatus";

export interface Room {
  code: string;
  status: RoomStatus;
  hostName: string;
  guestName?: string | null;
  questionCount: number;
  customQuestions?: CustomQuestion[];
  createdAt: Date;
}